# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@diana-health.com
    name: DIANA Support
  description: Diabetes Insulin Activity and Nutrition Analyzer (DIANA) - Risk assessment
    and patient management API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DIANA API
  version: "1.0"
paths:
...
  /admin/clinic-comparison:
    get:
      description: Returns statistics comparing all clinics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get clinic comparison (admin only)
      tags:
      - Admin
  /admin/clinics:
    get:
      description: Returns all clinics in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all clinics (admin only)
      tags:
      - Admin
  /admin/dashboard:
    get:
      description: Returns system-wide statistics (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get admin dashboard
      tags:
      - Admin
...
  /clinics:
    get:
      description: Returns all clinics the current user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List user's clinics
      tags:
      - Clinics
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"